Updated for changes to directx-vs-templates

+++ SimpleSampleWindows10/pch.h
@@ -11,7 +11,7 @@
 #include <wrl.h>
 
 #include <d3d11_3.h>
-#include <dxgi1_4.h>
+#include <dxgi1_6.h>
 #include <DirectXMath.h>
 #include <DirectXColors.h>
 
@@ -53,7 +53,7 @@ namespace DX
         virtual const char* what() const override
         {
             static char s_str[64] = {};
-            sprintf_s(s_str, "Failure with HRESULT of %08X", result);
+            sprintf_s(s_str, "Failure with HRESULT of %08X", static_cast<unsigned int>(result));
             return s_str;
         }
